+++ apps/e2e/specs/home.spec.ts
@@ -1,8 +1,0 @@
-import { expect, test } from "@playwright/test";
-import { HomePage } from "../page-objects/page";
-
-test("has title", async ({ page }) => {
-  const homePage = new HomePage(page);
-  await homePage.goto();
-  await expect(await homePage.getHeading()).toBeVisible();
-});